Makevars.win: Switch to "Biarch: yes"
Seems that win-builder (and thus the pre-submit checks) check with
--force-multiarch enabled. However, unless Biarch is turned on, CMD check will
install for one architecture (x64) but run tests on both (i386/x64). Then act
surprised why it doesn't work.

Also, "Biarch: yes" does nothing if you are using a Makefile.win, the option
only works on a "Makevars.win".

The reason for not using Makevars.win was that install.R will get SHLIB to
compile "*.cc" as well as the target we actually want. However, as the paramin
sources are in src-para now, this isn't a problem.