userguide: Replace colon with - in chap/section IDs
Prior to bookdown 0.23, HTML pages had : replaced with - in page
titles. This has now stopped working, and worse, gitbook output doesn't
escape HTML properly when inserting links, resulting in a malformed

  <a href="chap:input.html">

This is certainly a bug in bookdown, the colon should be escaped.
However, even if it was, our existing URLs break.

Instead, rename the IDs to use the dash, skirting around the problem
and keeping URLs working.